// packages/ui/src/recoil/tryout/createAccount/types.ts
export type AtomCreateAccountState = {
  accountBalanceGGEZ: string | null;
  accountBalanceGGEZOne: string | null;
  accountGGEZDenom: string | null;
  accountGGEZOneDenom: string | null;
  accountMnemonic: string | null;
  accountAddress: string | null;
  accountName: string | null;
  accountNumber: string | null;
  accountPublicKey: string | null;
};